backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pymongo
pydantic
python-dotenv
openai